refactor: change files folder structure
